requirements.txt
Bootstrap_Flask==2.2.0
Requests==2.31.0
Werkzeug==3.0.0
Flask==2.3.2
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
beautifulsoup4==4.12.3
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9